fix issues found during code review

+++ src/bacnet/basic/service/h_upt.c
@@ -49,7 +49,7 @@ void private_transfer_print_data(BACNET_PRIVATE_TRANSFER_DATA *private_data)
         application_data_len = private_data->serviceParametersLen;
         for (;;) {
             len = bacapp_decode_application_data(
-                application_data, (uint8_t)application_data_len, &value);
+                application_data, application_data_len, &value);
             if (first_value && (len < application_data_len)) {
                 first_value = false;
 #if PRINT_ENABLED
